Hoist Nobles class out to its own class file
