+++ PyRoute/Star.py
@@ -39,55 +39,6 @@ class UWPCodes(object):
             self.codes[uwpCode] = "X"
 
 
-class Nobles(object):
-    def __init__(self):
-        self.nobles = {'Knights': 0,
-                       'Baronets': 0,
-                       'Barons': 0,
-                       'Marquis': 0,
-                       'Viscounts': 0,
-                       'Counts': 0,
-                       'Dukes': 0,
-                       'Sector Dukes': 0,
-                       'Archdukes': 0,
-                       'Emperor': 0}
-        self.codes = {'B': 'Knights',
-                      'c': 'Baronets',
-                      'C': 'Barons',
-                      'D': 'Marquis',
-                      'e': 'Viscounts',
-                      'E': 'Counts',
-                      'f': 'Dukes',
-                      'F': 'Sector Dukes',
-                      'G': 'Archdukes',
-                      'H': 'Emperor'}
-
-    def __str__(self):
-        # If there's absolutely no nobles, return '-':
-        if 0 == max(self.nobles.values()):
-            return '-'
-
-        nobility = ""
-        for rank, count in self.nobles.items():
-            if count > 0:
-                nobility += list(self.codes.keys())[list(self.codes.values()).index(rank)]
-        return ''.join(sorted(nobility, key=lambda v: (v.lower(), v[0].isupper())))
-
-    def __getstate__(self):
-        state = self.__dict__.copy()
-        del state['codes']
-        return state
-
-    def count(self, nobility):
-        for code, rank in self.codes.items():
-            if code in nobility:
-                self.nobles[rank] += 1
-
-    def accumulate(self, nobles):
-        for rank, count in nobles.nobles.items():
-            self.nobles[rank] += count
-
-
 class Star(object):
     __slots__ = '__dict__', '_hash', '_key', 'index', 'zone', 'tradeCode', 'wtn', 'alg_code', 'hex'
